Make ZipappExtension an interface and add some @Deprecated annotations
